pyproject.toml
[tool.poetry]
name = "library"
version = "0.1.0"
description = ""
authors = ["ndestrieux <n.destrieux@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "~0.111"
strawberry-graphql = {extras = ["fastapi"], version = "~0.234"}
sqlalchemy = "~2.0"
uvicorn = {extras = ["standard"], version = "~0.30"}
pydantic-settings = "~2.2"


[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
pre-commit = "~3.7"
pytest = "~8.2"
pytest-asyncio = "^0.23.8"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.plugins.dotenv]
ignore = "false"
location = "./src/.env"


[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v -ra -q"
minversion = "6.0"
filterwarnings = ["ignore:'typing.ByteString' is deprecated and slated for removal in Python 3.14"]
